deepen: inline ai/parsers/normalizers.ts into response-parser.ts
Replaces 1 shallow extracted-for-testability module with private functions
inside its sole caller, the deep Response Parser module.

Before: lib/ai/parsers/normalizers.ts exported normalizeDate / normalizeCurrency
/ normalizePercentage (plus a private getMonthIndex helper). The functions had
ONE non-test caller — lib/ai/parsers/response-parser.ts, in the same package —
and ZERO dedicated test files. They were re-exported through the parsers
barrel, but no caller outside this module imported them through that path
either. Classic extracted-for-testability pattern: implementation detail
hoisted into its own module purely to be mockable in isolation, with no
locality or leverage gain.

After: the four functions are private implementation inside response-parser.ts.
The parsers barrel (lib/ai/parsers/index.ts) drops its `export * from
'./normalizers'` line — nothing outside the package imported through it.

Same shape as ADR-0002 (inline-analysis-depth-scoring): the deletion test
applies cleanly. Deleting normalizers.ts concentrates complexity at ONE
existing caller, not spreading it across N. Tests at the deepened module's
interface — lib/ai/parsers/response-parser.test.ts (parseResponse) and
__tests__/email/form4-field-normalizer.test.ts — already cover the
normalization behaviour observably, through parseResponse's filing-data output.
No tests deleted, no tests added.

Dependency category (per process/Deepening/deepening.md): in-process, pure
functions. No adapter needed.

Test sweep on lib/ai/parsers/ (response-parser, simple-parser,
x-sentiment-validator, etc.) shows IDENTICAL pre-existing failure counts on
main vs this branch (9 failed, 259 passed on both — failures unrelated to
this PR). No regressions.

Respects ADR-0002 (same pattern, same justification). No new ADR — this is
an instance of the pattern ADR-0002 already records.
